Avoid double callback calling
No need to handle `close` event

From `exec` docu:
On error, error will be an instance of Error. The error.code property will be the exit code of the child process while error.signal will   be set to the signal that terminated the process. Any exit code other than 0 is considered to be an error.

Fix #8